+++ geompp/src/mesh2d.cpp
@@ -63,12 +63,7 @@ PolyMesh2D Mesh2D::Polygonize(PolygonizationParams const& params) const {
   }
 
   auto pieces = detail::polygonize_impl(faces, params);
-
-  std::vector<Polygon2D> polygons;
-  polygons.reserve(pieces.size());
-  for (auto& [outer, holes] : pieces) {
-    polygons.push_back(Polygon2D::Make(std::move(outer), std::move(holes)));
-  }
+  auto polygons = detail::polygons_from_pieces(std::move(pieces));
   return PolyMesh2D::FromPolygons(polygons);
 }